UploadFirmware Complete
Script to upload firmware to a blank unit is functional on my end.  A blank unit can be plugged in and with a double click the script will check and make sure the required software is installed, then set the proper fuses, then upload the HEX provided.  HEX can be grabbed from a verbose arduino compile, but once the software is compiled it should only need to be done once.